@@ -3,11 +3,11 @@
     "description": "design patterns by php",
     "type": "project",
     "require": {
-        "php": ">=5.6.0"
+        "php": ">=7.0.0"
     },
     "require-dev": {
         "satooshi/php-coveralls": "*",
-        "phpunit/phpunit": "^5.7"
+        "phpunit/phpunit": "^6.3"
     },
     "autoload": {
         "psr-4": {
@@ -17,7 +17,7 @@
     "license": "MIT",
     "authors": [
         {
-            "name": "JShadowMan",
+            "name": "Wang",
             "email": "shadowman@shellboot.com"
         }
     ]
